Organize components into individual folders
Components may require additional files, such as tests, stories, or a readme, in order to be thoroughly defined. Consequently, a folder structure that is easy to use no matter the scale of a project should be established.
