+++ source/components/accordion/accordion.scss
@@ -1,8 +1,8 @@
 // ACCORDION
 // -----------------------------------------------------------------------------
 
-@use "../settings" as *;
-@use "../helpers/style-focus" as *;
+@use "../../settings" as *;
+@use "../../helpers/style-focus" as *;
 
 .accordion {
   &-segment {
