+++ source/components/badge/badge.scss
@@ -1,7 +1,7 @@
 // BADGE
 // -----------------------------------------------------------------------------
 
-@use "../settings" as *;
+@use "../../settings" as *;
 
 .badge {
   display: inline-flex;
