+++ source/components/breadcrumb/breadcrumb.scss
@@ -1,8 +1,8 @@
 // BREADCRUMB
 // -----------------------------------------------------------------------------
 
-@use "../settings" as *;
-@use "../helpers/style-focus" as *;
+@use "../../settings" as *;
+@use "../../helpers/style-focus" as *;
 
 .breadcrumb {
   display: flex;
