+++ source/components/button/button.scss
@@ -1,8 +1,8 @@
 // BUTTON
 // -----------------------------------------------------------------------------
 
-@use "../settings" as *;
-@use "../helpers/style-focus" as *;
+@use "../../settings" as *;
+@use "../../helpers/style-focus" as *;
 
 .button {
   display: inline-flex;
