+++ source/components/card/card.scss
@@ -1,8 +1,8 @@
 // CARD
 // -----------------------------------------------------------------------------
 
-@use "../settings" as *;
-@use "../helpers/style-focus" as *;
+@use "../../settings" as *;
+@use "../../helpers/style-focus" as *;
 
 .card {
   display: flex;
