+++ source/components/checkbox/checkbox.scss
@@ -1,8 +1,8 @@
 // CHECKBOX
 // -----------------------------------------------------------------------------
 
-@use "../settings" as *;
-@use "../helpers/style-focus" as *;
+@use "../../settings" as *;
+@use "../../helpers/style-focus" as *;
 
 .checkbox {
   width: 1.5rem;
