+++ source/components/global-header/global-header.scss
@@ -1,9 +1,9 @@
 // GLOBAL HEADER
 // -----------------------------------------------------------------------------
 
-@use "../settings" as *;
-@use "../helpers/style-focus" as *;
-@use "../helpers/set-breakpoint" as *;
+@use "../../settings" as *;
+@use "../../helpers/style-focus" as *;
+@use "../../helpers/set-breakpoint" as *;
 
 .global-header {
   &-container {
