+++ source/components/link/link.scss
@@ -1,8 +1,8 @@
 // LINK
 // -----------------------------------------------------------------------------
 
-@use "../settings" as *;
-@use "../helpers/style-focus" as *;
+@use "../../settings" as *;
+@use "../../helpers/style-focus" as *;
 
 .link {
   display: inline-flex;
