+++ source/components/menu/menu.scss
@@ -1,9 +1,9 @@
 // MENU
 // -----------------------------------------------------------------------------
 
-@use "../settings" as *;
-@use "../helpers/animate-phasing" as *;
-@use "../helpers/style-focus" as *;
+@use "../../settings" as *;
+@use "../../helpers/animate-phasing" as *;
+@use "../../helpers/style-focus" as *;
 
 .menu {
   position: absolute;
