+++ source/components/navigation/navigation.scss
@@ -1,8 +1,8 @@
 // NAVIGATION
 // -----------------------------------------------------------------------------
 
-@use "../settings" as *;
-@use "../helpers/style-focus" as *;
+@use "../../settings" as *;
+@use "../../helpers/style-focus" as *;
 
 .navigation {
   display: flex;
