+++ source/components/notice/notice.scss
@@ -1,7 +1,7 @@
 // NOTICE
 // -----------------------------------------------------------------------------
 
-@use "../settings" as *;
+@use "../../settings" as *;
 
 .notice {
   padding: $space-1 $space-1-5;
