+++ source/components/radio/radio.scss
@@ -1,23 +1,19 @@
-// SWITCH
+// RADIO
 // -----------------------------------------------------------------------------
 
-@use "../settings" as *;
-@use "../helpers/style-focus" as *;
+@use "../../settings" as *;
+@use "../../helpers/style-focus" as *;
 
-.switch {
-  width: 2.5rem;
+.radio {
+  width: 1.5rem;
   height: 1.5rem;
   border: $border-width-1 solid $color-neutral-250;
   border-radius: $border-radius-m;
   vertical-align: middle;
-  background-image: url("data:image/svg+xml;utf8,<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'><circle cx='8' cy='8' r='5' fill='%233d3d3d'/></svg>");
-  background-position: left center;
-  background-size: 1.5rem;
-  background-repeat: no-repeat;
   background-color: $color-white;
   cursor: pointer;
   -webkit-appearance: none;
-  transition: border-color $transition-ease-out-100, background-position $transition-ease-out-100, background-color $transition-ease-out-100, outline $transition-ease-out-100;
+  transition: border-color $transition-ease-out-100, background-color $transition-ease-out-100, outline $transition-ease-out-100;
 
   &:focus-visible {
     @include style-focus;
@@ -26,7 +22,9 @@
   &:checked {
     border-color: $color-brand-100;
     background-image: url("data:image/svg+xml;utf8,<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'><circle cx='8' cy='8' r='5' fill='%23fff'/></svg>");
-    background-position: right center;
+    background-position: center;
+    background-size: 1rem;
+    background-repeat: no-repeat;
     background-color: $color-brand-100;
   }
 
@@ -37,7 +35,7 @@
 
 // Label
 
-.switch-label {
+.radio-label {
   padding-left: $space-0-25;
   vertical-align: middle;
   cursor: pointer;
