+++ source/components/table/table.scss
@@ -1,7 +1,7 @@
 // TABLE
 // -----------------------------------------------------------------------------
 
-@use "../settings" as *;
+@use "../../settings" as *;
 
 .table {
   width: 100%;
