+++ source/components/tabset/tabset.scss
@@ -1,9 +1,9 @@
 // TABSET
 // -----------------------------------------------------------------------------
 
-@use "../settings" as *;
-@use "../helpers/animate-phasing" as *;
-@use "../helpers/style-focus" as *;
+@use "../../settings" as *;
+@use "../../helpers/animate-phasing" as *;
+@use "../../helpers/style-focus" as *;
 
 .tabset {
   &-navigation {
