+++ source/components/textfield/textfield.scss
@@ -1,8 +1,8 @@
 // TEXTFIELD
 // -----------------------------------------------------------------------------
 
-@use "../settings" as *;
-@use "../helpers/style-focus" as *;
+@use "../../settings" as *;
+@use "../../helpers/style-focus" as *;
 
 .textfield {
   display: block;
