+++ source/main.scss
@@ -6,25 +6,25 @@
 // COMPONENTS
 // -----------------------------------------------------------------------------
 
-@use "./components/accordion";
-@use "./components/advisory";
-@use "./components/badge";
-@use "./components/breadcrumb";
-@use "./components/button";
-@use "./components/card";
-@use "./components/checkbox";
-@use "./components/dialog";
-@use "./components/global-footer";
-@use "./components/global-header";
-@use "./components/link";
-@use "./components/menu";
-@use "./components/navigation";
-@use "./components/notice";
-@use "./components/radio";
-@use "./components/switch";
-@use "./components/table";
-@use "./components/tabset";
-@use "./components/textfield";
+@use "./components/accordion/accordion";
+@use "./components/advisory/advisory";
+@use "./components/badge/badge";
+@use "./components/breadcrumb/breadcrumb";
+@use "./components/button/button";
+@use "./components/card/card";
+@use "./components/checkbox/checkbox";
+@use "./components/dialog/dialog";
+@use "./components/global-footer/global-footer";
+@use "./components/global-header/global-header";
+@use "./components/link/link";
+@use "./components/menu/menu";
+@use "./components/navigation/navigation";
+@use "./components/notice/notice";
+@use "./components/radio/radio";
+@use "./components/switch/switch";
+@use "./components/table/table";
+@use "./components/tabset/tabset";
+@use "./components/textfield/textfield";
 
 // LAYOUTS
 // -----------------------------------------------------------------------------
